+++ src/game_version.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const game_version = "v0.6.12";
+const game_version = "v0.6.13";
 
 function get_game_version() {
     return game_version;